requirements.txt
selenium==4.23.1
webdriver-manager==4.0.2
PyYAML==6.0.2
pytest==8.3.2

